# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  aslon1213_gift_pkg_repository.Empty:
    type: object
...
  aslon1213_gift_pkg_repository.Response-any:
    properties:
      data: {}
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
...
  aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty:
    properties:
      data:
        $ref: '#/definitions/aslon1213_gift_pkg_repository.Empty'
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
...
  aslon1213_gift_pkg_repository.Response-pkg_handlers_SettingsResponse:
    properties:
      data:
        $ref: '#/definitions/pkg_handlers.SettingsResponse'
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
...
  pkg_handlers.ProfileInfo:
    properties:
      balance:
        type: number
      currency:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  pkg_handlers.ServerInfo:
    properties:
      host:
        type: string
      online:
        type: boolean
      started_at:
        type: string
      uptime:
        type: string
      uptime_seconds:
        type: integer
      version:
        type: string
    type: object
  pkg_handlers.ServerStats:
    properties:
      budgets:
        type: integer
      db_size_bytes:
        type: integer
      goals:
        type: integer
      goroutines:
        type: integer
      groups:
        type: integer
      mem_mb:
        type: integer
      users:
        type: integer
    type: object
  pkg_handlers.SettingsResponse:
    properties:
      profile:
        $ref: '#/definitions/pkg_handlers.ProfileInfo'
      server:
        $ref: '#/definitions/pkg_handlers.ServerInfo'
      stats:
        $ref: '#/definitions/pkg_handlers.ServerStats'
    type: object
...
host: localhost:3000
info:
  contact:
    email: hamidovaslon1@gmail.com
    name: API Support
    url: https://github.com/aslon1213/gift
  description: This is a Gift API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/aslon1213/gift
  title: Gift API
  version: 0.1.0
paths:
...
  /api/v1/settings:
    get:
      description: Returns server identity (host/uptime/version), runtime stats and
        the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-pkg_handlers_SettingsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: Get settings snapshot
      tags:
      - settings
  /api/v1/settings/export_data:
    post:
      description: Export all collections data from the database in JSON or CSV format
      parameters:
      - default: json
        description: 'Export format: json or csv'
        enum:
        - json
        - csv
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: Export all data
      tags:
      - settings
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"